rename CourseRest files, resolve unused variables
