+++ src/components/CoursesRest/CourseSingleton.js
