+++ src/components/CoursesRest/index.js
@@ -1,4 +1,4 @@
-export { default as CourseSingleton } from './CourseCardLink';
-export { default as CourseList } from './CourseListLink';
-export { default as AddCourseForm } from './REST_AddCourseForm';
-export { default as EditCourseForm } from './REST_EditCourseForm';
+export { default as CourseSingleton } from './CourseSingleton';
+export { default as CourseList } from './CourseList';
+export { default as AddCourseForm } from './AddCourseForm';
+export { default as EditCourseForm } from './EditCourseForm';
